+++ BraintreeCore/src/main/java/com/braintreepayments/api/BraintreeClient.kt
@@ -225,9 +225,9 @@ open class BraintreeClient @VisibleForTesting internal constructor(
         configuration: Configuration?,
         authorization: Authorization
     ) {
-        if (isAnalyticsEnabled(configuration)) {
+        configuration?.let {
             analyticsClient.sendEvent(
-                configuration!!,
+                it,
                 eventName,
                 sessionId,
                 integrationType,
@@ -429,7 +429,15 @@ open class BraintreeClient @VisibleForTesting internal constructor(
      */
     @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
     fun reportCrash() = authorizationLoader.authorizationFromCache?.let { authorization ->
-        analyticsClient.reportCrash(applicationContext, sessionId, integrationType, authorization)
+        getConfiguration { configuration, _ ->
+            analyticsClient.reportCrash(
+                applicationContext,
+                configuration,
+                sessionId,
+                integrationType,
+                authorization
+            )
+        }
     }
 
     /**
@@ -462,16 +470,5 @@ open class BraintreeClient @VisibleForTesting internal constructor(
      */
     open fun launchesBrowserSwitchAsNewTask(launchesBrowserSwitchAsNewTask: Boolean) {
         this.launchesBrowserSwitchAsNewTask = launchesBrowserSwitchAsNewTask
-    }
-
-    companion object {
-
-        /**
-         * @suppress
-         */
-        @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-        fun isAnalyticsEnabled(configuration: Configuration?): Boolean {
-            return configuration != null && configuration.isAnalyticsEnabled
-        }
     }
 }